requirements.txt
Django>=5.0
psycopg2-binary>=2.9
requests>=2.31
celery>=5.3
redis>=5.0
django-environ>=0.11
python-decouple>=3.8
dj-database-url>=2.1.0
django-htmx>=1.17.0